requirements.txt
alembic==1.10.3
aniso8601==9.0.1
attrs==23.1.0
Authlib==1.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.2
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
requests==2.29.0
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
